Viscoacoustoelasticity

This package provides the tools to compute the dispersion relation of small amplitude waves in homogeneously pre-stressed viscoelastic media. In the framework of the acoustoelatic theory, the propagation of incremental waves on top of a finitely deformed configuration is still described by a wave equation where the elasticity tensor $\boldsymbol{C}$ is replaced by a modified elasticity tensor $\boldsymbol{C^0}$, which takes into account the intertwined contributions of pre-stress and viscoleasticity, and where $\boldsymbol{u}$ now stands for the incremental displacement.

This similarity allows to use the methods introduced above for elastic waveguides to compute the dispersion relation in pre-stressed viscoelastic media after:

computing the modified elasticity tensor $\boldsymbol{C^0}$

adapting the geometry to the considered level of pre-stress, for example $w = \lambda_\theta w_0$, and $h = h_0/(\lambda_\theta \lambda_x)$.

We include functions to obtain the 4-th order incremental $\boldsymbol{C^0}$ for nearly incompressible hyperelastic materials. $\boldsymbol{C^0}$ is computed from a chosen strain energy density function $W$, built on the principal invariants of the left Cauchy-Green tensor $\boldsymbol{B} = \boldsymbol{F}\cdot\boldsymbol{F}^T$,

\[\begin{align*} &I_1 = tr(\boldsymbol{B}) = \lambda_1^2 + \lambda_2^2 +\lambda_3^2, \\ &I_2 = \frac{1}{2}\left(\mathrm{Tr}(\boldsymbol{B})^2 - \mathrm{Tr}(\boldsymbol{B}^2)\right) = \lambda_2^2\lambda_3^2 + \lambda_1^2\lambda_3^2 + \lambda_1^2\lambda_2^2, \\ &I_3 = \mathrm{det}(\boldsymbol{B}) = \lambda_1^2\lambda_2^2\lambda_3^2 = J^2. \end{align*}\]

The coefficients of the push-forward of $\boldsymbol{C}$ in the deformed configuration can be found in Delory et al., 2024 and are identical to that given in Ogden, 1997 with a permutation of the last two indices. The effect of viscoelasticity is incorporated in the modified elasticity tensor $\boldsymbol{C^0}$, by adding a term to the push-forward of $\boldsymbol{C}$ based on the fractional Kelvin-Voigt model. $\boldsymbol{C^0}$ becomes frequency dependent and reads

\[\begin{equation*} C^0_{jikl}(\lambda_1,\lambda_2,\lambda_3, \omega) = C^0_{jikl}(\lambda_1,\lambda_2,\lambda_3)+\mu_0I_{jikl}\left(1 + \beta'\frac{\lambda_i^2+\lambda_j^2-2}{2}\right)(\mathrm{i}\omega\tau)^n, \end{equation*}\]

with $I_{jikl} = (\delta_{jk}\delta_{il}+\delta_{jl}\delta_{ik})$ as detailed in Delory et al., 2023. The second term, that involves both the principal stretches and the frequency, underlines the coupling between pre-stress and viscoelasticity.

The following constitutive relations have been implemented where we further remove the dependence on the stretch ratio $\lambda_2$ by taking the nearly incompressible limit $J \to 1$ in combination with the boundary condition $\sigma_2 = 0$.

Tip

If you want to implement a different elasticity tensor (i.e. different hyperelastic model, viscoelastic model) have a look at the Mathematica files provided in the GEW soft strip repository.

Models for the small to moderate strain regime

POnG.C_MR — Function

C_MR(lambda1,lambda3,rho,vL,C1,C2,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Mooney-Rivlin model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Mooney-Rivlin strain energy density function is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3\right) + \frac{1}{2} C_2 \left(\frac{I_2}{J^{4/3}} - 3\right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_GT — Function

C_GT(lambda1,lambda3,rho,vL,C1,C2,f,tau,n,betaprime)

Returns the 4-th order elastic tensor for a nearly incompressible hyperelastic medium described by the Gent-Thomas model for a large deformation specified through the elongations in two principal directions lambda1 and lambda3. The stretch ratio in the third direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Gent-Thomas strain energy density function is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3 \right) + \frac{3}{2} C_2 ln \left(\frac{I_2}{3J^{4/3}}\right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_C — Function

C_C(lambda1,lambda3,rho,vL,C1,C2,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Carroll model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Carroll strain energy density function is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3\right) + \sqrt{3} C_2 \left(\sqrt{\frac{I_2}{J^{4/3}}} - \sqrt{3}\right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

Models for the strain-hardening regime

POnG.C_MRSH — Function

C_MRSH(lambda1,lambda3,rho,vL,C1,C2,C3,N,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Mooney-Rivlin model with a $I_1$strain hardening term subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Mooney-Rivlin strain energy density function with a $I_1$ strain hardening term is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3\right) + \frac{1}{2} C_2 \left(\frac{I_2}{J^{4/3}} - 3\right) + \frac{3^{1-N}}{2N} C_3 \left(\left(\frac{I_1}{J^{2/3}}\right)^N - 3^N \right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3, N: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_MRSH2 — Function

C_MRSH2(lambda1,lambda3,rho,vL,C1,C2,C3,N,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Mooney-Rivlin model with a $I_2$ strain hardening term subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Mooney-Rivlin strain energy density function with a $I_2$ strain hardening term is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3\right) + \frac{1}{2} C_2 \left(\frac{I_2}{J^{4/3}} - 3\right) + \frac{3^{1-N}}{2N} C_3 \left(\left(\frac{I_2}{J^{4/3}}\right)^N - 3^N \right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3, N: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_GTSH — Function

C_GTSH(lambda1,lambda3,rho,vL,C1,C2,C3,N,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Gent-Thomas model with a $I_1$ strain hardening term subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Gent-Carroll strain energy density function is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3 \right) + \frac{3}{2} C_2 ln \left(\frac{I_2}{3J^{4/3}}\right) + \frac{3^{1-N}}{2N} C_3 \left(\left(\frac{I_1}{J^{2/3}}\right)^N - 3^N \right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3,N: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_GTSH2 — Function

C_GTSH2(lambda1,lambda3,rho,vL,C1,C2,C3,N,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Gent-Thomas model with a $I_2$ strain hardening term subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Mooney-Rivlin strain energy density function with a $I_2$ strain hardening term is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3\right) + \frac{3}{2} C_2 ln \left(\frac{I_2}{3J^{4/3}}\right) + \frac{3^{1-N}}{2N} C_3 \left(\left(\frac{I_2}{J^{4/3}}\right)^N - 3^N \right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3, N: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_CSH — Function

C_CSH(lambda1,lambda3,rho,vL,C1,C2,C3,N,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Carroll model with a$I_1$ strain hardening term subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Gent-Carroll strain energy density function is writen as $W = \frac{1}{2} C_1 \left(\frac{I_1}{J^{2/3}} - 3 \right) + \sqrt{3} C_2 \left(\sqrt{\frac{I_2}{J^{4/3}}} - \sqrt{3}\right) + \frac{3^{1-N}}{2N} C_3 \left(\left(\frac{I_1}{J^{2/3}}\right)^N - 3^N \right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3,N: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

Models for the limiting-chain regime

POnG.C_GMR — Function

C_GMR(lambda1,lambda3,rho,vL,C1,C2,C3,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Gent-Mooney-Rivlin model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Gent-Mooney-Rivlin strain energy density function is writen as $W = - \frac{1}{2} C_1 C_3 ln \left(1 - \frac{I_1 / J^{2/3} - 3}{C_3}\right) + \frac{1}{2} C_2 \left(\frac{I_2}{J^{4/3}} - 3\right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_GG — Function

C_GG(lambda1,lambda3,rho,vL,C1,C2,C3,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Gent-Gent model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Gent-Gent strain energy density function is writen as $W = - \frac{1}{2} C_1 C_3 ln \left(1 - \frac{I_1 / J^{2/3} - 3}{C_3}\right) + \frac{3}{2} C_2 ln\left(\frac{I_2}{3 J^{4/3}}\right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_GC — Function

C_GC(lambda1,lambda3,rho,vL,C1,C2,C3,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Gent-Carroll model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Gent-Carroll strain energy density function is writen as $W = - \frac{1}{2} C_1 C_3 ln \left(1 - \frac{I_1 / J^{2/3} - 3}{C_3}\right) + \sqrt{3} C_2 \left(\sqrt{\frac{I_2}{J^{4/3}}} - \sqrt{3}\right) + \frac{K}{2} (J - 1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_DCMR2 — Function

C_DCMR2(lambda1,lambda3,rho,vL,C1,C2,C3,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Dobrynin & Carrillo-Mooney-Rivlin model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Dobrynin & Carrillo-Mooney-Rivlin strain energy density function is writen as $W = \frac{C_1}{6} \left( \frac{I_1}{J^{2/3}} - 3 + \frac{2C_3}{1 - \left(I_1/J^{2/3} - 3\right)/C_3} - 2C_3\right) + \frac{C_2}{2} \left(\frac{I_2}{J^{4/3}} - 3\right) + \frac{\kappa}{2}(J-1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity

POnG.C_DCGT — Function

C_DCGT(lambda1,lambda3,rho,vL,C1,C2,C3,f,tau,n,betaprime)

Returns the 4-th order incremental elastic tensor for a nearly incompressible hyperelastic medium described by the Dobrynin & Carrillo-Gent-Thomas model subjected to a large deformation specified by the elongations in the principal directions 1, lambda1, and 3, lambda3. The stretch ratio in the remaining direction is given by imposing the stress free condition in direction 2. We also consider viscoelastic material properties through the fractional Kelvin-Voigt model.

The Dobrynin & Carrillo-Mooney-Rivlin strain energy density function is writen as $W = \frac{C_1}{6} \left(\frac{I_1}{J^{2/3}} - 3 + \frac{2C_3}{1 - \left(I_1/J^{2/3} - 3\right)/C_3} - 2C_3\right) + \frac{3}{2} C_2 ln \left(\frac{I_2}{3J^{4/3}}\right) + \frac{\kappa}{2}(J-1)^2$, where $I_1$, $I_2$ and $J^2$ are the first three invariants of the left Cauchy-Green tensor.

Input arguments:

lambda1: stretch ratio in direction 1

lambda3: stretch ratio in direction 3

rho: material density

vL: longitudinal wave velocity

C1, C2, C3: hyperelastic parameters

f: frequency

tau: relaxation time

n: fractional derivative order

betaprime: coupling between viscoelasticity and hyperelasticity